requirements.txt
streamlit-pydantic
pydantic==1.10.14
psycopg2-binary==2.9.9
python-dotenv
pandas
requests
BeatifulSoup4